requirements.txt
wxPython>=4.2.1
python-vlc>=3.0.21203
accessible-output2>=0.17
